several sample functions & deployed
