+++ functions/package.json
@@ -10,11 +10,12 @@
     "logs": "firebase functions:log"
   },
   "engines": {
-    "node": "8"
+    "node": "10"
   },
   "main": "lib/index.js",
   "dependencies": {
     "cors": "^2.8.5",
+    "express": "^4.17.1",
     "express-jwt": "^5.3.3",
     "firebase-admin": "^8.10.0",
     "firebase-functions": "^3.6.1",
