+++ functions/src/admin/admin.ts
@@ -1,4 +1,4 @@
-import * as functions from 'firebase-functions';
+
 import * as admin from "firebase-admin";
 
 /**
